% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{constants}
\alias{constants}
\alias{file.drawer.path}
\alias{map.path}
\title{Constants.}
\format{Multiple formats}
\usage{
file.drawer.path

map.path
}
\description{
Frequently used constants.
}
\details{
\bold{Constants for ggmap}:\cr
\bold{file.drawer.path}: //tec-storage2/tec10189/Analytics/GIS/Maps/Stamen/ggmap\cr
\bold{map.path}: //Cir-Nas/Users/TEC10189/Analytics/GIS/Maps/
}
\author{
John A. Clouse
}
\keyword{datasets}

