% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ebi.logo.png}
\alias{ebi.logo.png}
\title{EBI Team Logo.}
\format{A png object}
\usage{
ebi.loglo.png
}
\description{
A 165 x 149 pixel PNG file of the EBI Team logo.  This logo is used as a
watermark in plot outputs and for branding in documents.
}
\details{
Logo can be rendered using the grid.package e.g.,
}
\examples{
library(grid)
grid.raster(ebi.logo.png)
}

