% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/splitCapWords.R
\name{splitCapWords}
\alias{splitCapWords}
\title{Split Words with Medial Symbols}
\usage{
splitCapWords(x)
}
\arguments{
\item{x}{a string}
}
\value{
string with words split on medial symbols
}
\description{
Split words containing medial capital letters, periods, or underscores
}
\details{
For each word within a string, split words with medial symbols (capitals,
periods, or underscores) by addinga a space before all medial symbols
preceeded by a lower case letter
}
\examples{
splitCapWords("TestingThisString")
splitCapWords("Testing.this.String")
splitCapWords("testing.this.string")
splitCapWords("testing_this.stRing")
}
\author{
John A. Clouse
}
\seealso{
\code{\link{gsub}}
}

