% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/toTitle.R
\name{toTitle}
\alias{toTitle}
\title{Convert a String to Psuedo Title Case}
\usage{
toTitle(x)
}
\arguments{
\item{x}{a string}
}
\value{
string converted to psuedo title case
}
\description{
Convert a string to psuedo title case
}
\details{
Converts all characters to lower case and then changes the
first character of each word to be upper case.
}
\examples{
# Sample data
x <- "convert.this to titleCase and This"

# Note: title case is not proper but rather first letter capitalized case
# Note: capitals appearing mid-word will be converted to lower case
toTitle(x)
}
\author{
John A. Clouse
}
\seealso{
\code{\link{toupper}}
}

