% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vectorToTable.R
\name{vectorToTable}
\alias{vectorToTable}
\title{Split Vector Into Multiple Columns of Data Frame}
\usage{
vectorToTable(vector.to.split, width = 3, by.row = FALSE)
}
\arguments{
\item{original.vector,}{number.of.colums}
}
\value{
a data frame containing x split on medial symbols
}
\description{
Split a vector into a data frame of n columns for use with xtabl
}
\details{
A vector too long for printing as a single column table is split into n columns.  The original vector is
split into number.of.columns.  Any extra cells are filled with "".  The resulting table
is sorted by columns if by.row is left at the default value of FALSE.
}
\examples{
sample.vector <- 1:100
vectorToTable(sample.vector, 5)

sample.vector <- sort(replicate(23, paste(sample(LETTERS, 5, replace=TRUE), collapse="")))
vectorToTable(sample.vector, 5)
vectorToTable(sample.vector, 5, by.row = TRUE)
}
\author{
John A. Clouse
}
\seealso{
\code{\link{gsub}}
}

