% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/view.R
\name{view}
\alias{view}
\title{View Data Frame in Excel}
\usage{
view(x)
}
\arguments{
\item{x}{a data frame}
}
\description{
View data frame in Excel
}
\details{
View an existing data frame in an Excel workbook.  Row numbers are included.
Requires the existance of Microsoft Excel or another program associated with
the .csv file extension.

Note: the data frame is written to the \code{tempdir()}.  The file will be
left behind to be cleaned up by the user or the operating system (not
guaranteed.)
}
\examples{
x <- 1
}
\author{
John A. Clouse
}

